backend_py/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
pydantic==2.7.4
boto3==1.34.135
python-jose[cryptography]==3.3.0
requests==2.32.3
python-dotenv==1.0.1
uuid==1.30
